# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Horizon Pulse",
    "description": "Pay-per-call APIs for AI agents via x402 on Base: web fetch, HTTP proxy, page extract, and crypto market data. Eleven x402 v2 routes (USDC on Base mainnet, eip155:8453; Coinbase CDP facilitator): /api/fetch ($0.02 clean text/markdown), /api/http ($0.01 SSRF-safe HTTP proxy, GET|POST), /api/extract ($0.015 structured page fields, GET|POST), /api/x402-check ($0.01 x402 endpoint audit), /api/screenshot ($0.02 headless Chromium render), and six crypto routes (pulse, signals, yield, portfolio, gas, funding).",
    "version": "1.0.0",
    "contact": {
      "name": "Horizon Pulse",
      "email": "horizonpulse.co@proton.me",
      "url": "https://horizonpulse.dev"
    }
  },
  "servers": [
    {
      "url": "https://horizonpulse.dev",
      "description": "Canonical custom domain"
    },
    {
      "url": "https://horizon-pulse-seven.vercel.app",
      "description": "Vercel backup / fallback"
    }
  ],
  "paths": {
    "/api/pulse": {
      "get": {
        "summary": "BTC/ETH/SOL spot + momentum (CoinGecko)",
        "description": "x402 v2 exact payment. Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.005 USDC (5000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9.",
        "responses": {
          "200": {
            "description": "Paid success \u2014 live pulse JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          }
        }
      }
    },
    "/api/signals": {
      "get": {
        "summary": "RSI/MACD/Bollinger + OKX funding",
        "description": "x402 v2 exact payment. Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.015 USDC (15000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9.",
        "responses": {
          "200": {
            "description": "Paid success \u2014 live signals JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          }
        }
      }
    },
    "/api/yield": {
      "get": {
        "summary": "DefiLlama yields (TVL \u2265 $10M)",
        "description": "x402 v2 exact payment. Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.02 USDC (20000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9.",
        "responses": {
          "200": {
            "description": "Paid success \u2014 live yield JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          }
        }
      }
    },
    "/api/portfolio": {
      "get": {
        "summary": "Base + Ethereum portfolio risk snapshot",
        "description": "x402 v2 exact payment. Requires query param address=0x\u2026 Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.04 USDC (40000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9.",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$"
            },
            "description": "EVM address (0x + 40 hex)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid success \u2014 live portfolio JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          }
        }
      }
    },
    "/api/gas": {
      "get": {
        "summary": "Base + Ethereum gas fee snapshot",
        "description": "x402 v2 exact payment. Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.01 USDC (10000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9.",
        "responses": {
          "200": {
            "description": "Paid success \u2014 live gas JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          }
        }
      }
    },
    "/api/funding": {
      "get": {
        "summary": "OKX BTC/ETH/SOL perpetual funding",
        "description": "x402 v2 exact payment. Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.01 USDC (10000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9.",
        "responses": {
          "200": {
            "description": "Paid success \u2014 live funding JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          }
        }
      }
    },
    "/api/fetch": {
      "get": {
        "summary": "Fetch public URL \u2192 clean text/markdown",
        "description": "x402 v2 exact payment. Requires query param url=https://\u2026 Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.02 USDC (20000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9. SSRF-safe; ~200KB / 8s caps; preferred Content-Types: text/html, text/plain, application/json, text/markdown.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Absolute http(s) URL to fetch (private/localhost blocked)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid success \u2014 cleaned text/markdown JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          },
          "400": {
            "description": "Bad or blocked URL"
          },
          "413": {
            "description": "Upstream body too large"
          },
          "415": {
            "description": "Unsupported Content-Type"
          },
          "502": {
            "description": "Upstream error"
          },
          "504": {
            "description": "Upstream timeout"
          }
        }
      }
    },
    "/api/http": {
      "get": {
        "summary": "Universal HTTP proxy (query)",
        "description": "x402 v2 exact payment. Requires query param url=https://\u2026. Optional method (GET default), headers (JSON string). Unpaid \u2192 HTTP 402 + PAYMENT-REQUIRED. Price $0.01 USDC (10000 atomic). Network eip155:8453. payTo 0x5b32c973596078a967562ca652761404f19be0e9. SSRF-safe; ~384KB / 12s caps. Response: ok, status, headers, body, bodyEncoding, contentType, elapsedMs.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Absolute http(s) URL (private/localhost blocked)"
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "GET",
                "POST",
                "HEAD",
                "PUT",
                "PATCH",
                "DELETE"
              ],
              "default": "GET"
            }
          },
          {
            "name": "headers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional JSON object string of allowlisted outbound headers"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid success \u2014 proxied response JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          },
          "400": {
            "description": "Bad or blocked URL / method / headers"
          },
          "413": {
            "description": "Request or upstream body too large"
          },
          "502": {
            "description": "Upstream error"
          },
          "504": {
            "description": "Upstream timeout"
          }
        }
      },
      "post": {
        "summary": "Universal HTTP proxy (JSON body)",
        "description": "Same $0.01 USDC x402 price as GET. JSON body: { url, method?, headers?, body? }. body allowed for POST/PUT/PATCH upstream. Same SSRF/caps/payTo as GET.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "method": {
                    "type": "string",
                    "enum": [
                      "GET",
                      "POST",
                      "HEAD",
                      "PUT",
                      "PATCH",
                      "DELETE"
                    ]
                  },
                  "headers": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "body": {
                    "description": "String or JSON value (stringified) for POST/PUT/PATCH"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paid success \u2014 proxied response JSON"
          },
          "402": {
            "description": "Payment Required (x402 v2)"
          },
          "400": {
            "description": "Bad input / blocked target"
          },
          "413": {
            "description": "Body too large"
          },
          "502": {
            "description": "Upstream error"
          },
          "504": {
            "description": "Upstream timeout"
          }
        }
      }
    },
...
  }
}